Update README.md
This data is from NYC Open Data.  This is a 3 year record of all 311 calls to the Department of Sanitation that were created  from January 1, 2015 to December 31, 2017.
Kaggle Notebook for data analysis  is here => https://www.kaggle.com/anakyani/dsny-ana

And this repo contains Power BI report for this data with the visualizations.